[Build] mangosd services on IService; disarm acelite -- zero ACE in src/
CP5, the last checkpoint of Stage 2: mangosd's background services move
onto a shared IService interface (Name/Start/RequestStop/Join), and
acelite is dropped from the dependency graph entirely.

New Master class owns the databases, the world loop (now the main
thread's own loop, not a spawned WorldThread), and the service list --
StartDatabases/StopDatabases/ClearOnlineAccounts/StartServices/
StopServices/WorldLoop/ShutdownWorld/Run. AFThread/CliThread/RAThread/
WorldThread are deleted and replaced by AntiFreezeService, CliService,
RASession+RaServer+RaService, and Master's own WorldLoop:
  - RASession (net::ISession) carries RAThread.cpp's telnet login/
    password/command protocol verbatim (Stage/HandleUsername/
    HandlePassword/HandleCommand, the CliCommandHolder callback pair,
    the m_keepAlive self-reference while a command is pending).
  - CliService reads console commands off a std::thread with a POSIX
    select()-based non-blocking poll / Windows synthetic-Enter-key
    unblock on RequestStop().
  - AntiFreezeService is a std::condition_variable watchdog; it now
    calls std::abort() on a detected hang instead of the old raw
    *((uint32 volatile*)NULL) = 0 UB write -- a deliberate hardening,
    not a behavior-preserving port.
  - SoapService is a thin IService wrapper around the pre-existing
    (already ACE-free) SoapThread().
mangosd.cpp shrinks to hand-rolled argv parsing (matching realmd's
Main.cpp from the last commit) plus `Master().Run()`.

Dependency graph: root CMakeLists.txt's add_subdirectory(dep) becomes
include(cmake/Dependencies.cmake), which add_subdirectory()s every dep
submodule library EXCEPT acelite by name -- zlib, bzip2, utf8cpp,
lualib, gsoap, recastnavigation, g3dlite, StormLib, gated the same way
dep/CMakeLists.txt already gated them. dep/ (the mangosDeps submodule)
is untouched; acelite's sources are still there on disk, just never
added to the build. src/shared/CMakeLists.txt drops ace from
target_link_libraries and picks up Threads::Threads PUBLIC instead --
ACE was dragging pthread in as a side effect on Linux, a dependency
nothing here had declared until acelite's removal made it visible.
src/modules/Bots/CMakeLists.txt loses its ACE_USE_EXTERNAL dependency
hook (playerbots needs its own future ACE conversion, out of scope
here); src/game/BattleGround/CMakeLists.txt loses a stray orphaned
`ace` link token.

Common.h loses ACE_OFF_T and ace/config-all.h, its last two ACE
footprints. Removing config-all.h surfaced a real latent bug, not
a regression in this commit: on Windows, config-all.h was pulling in
<windows.h> with WIN32_LEAN_AND_MEAN defined ahead of it, which as a
side effect kept wincrypt.h out of the transitive include graph.
Without that, plain <windows.h> (via <ws2tcpip.h>) drags wincrypt.h in,
and its `#define CRYPT_FAILED FALSE` textually clobbers AuthCrypt.h's
CryptState enumerator of the same name -- caught by a --clean-first
rebuild of `shared`, not by incremental builds, which is why it didn't
surface at CP4. Fixed by defining WIN32_LEAN_AND_MEAN in Common.h
ourselves, matching the convention net/iocp/IocpServer.hpp already
uses. Two more ACE-transitive-include gaps turned up the same way, both
fixed with the minimal, targeted include rather than restoring
anything ACE: PosixDaemon.cpp needed <sys/stat.h> for umask() (Linux
only), and WorldSession.cpp needed <arpa/inet.h> (POSIX) / <winsock2.h>
+ <ws2tcpip.h> (Windows) for inet_addr() in SendRedirectClient() --
same pattern already established in realmd/Realm/RealmList.cpp.

Also fixed, unrelated to ACE: three GCC-only test files
(mop_client_opcode_worklist_test.cpp, mop_control_packets_test.cpp,
mop_movement_packets_test.cpp) called ObjectGuid(0x...ull) directly,
which is ambiguous on Linux between ObjectGuid(uint64) and the
intentionally-unimplemented ObjectGuid(uint32 const&) trap overload --
uint64_t is `unsigned long` on 64-bit Linux/GCC but `unsigned long
long` on MSVC, so an `...ull` literal is an equally-ranked conversion
to either parameter there and not on Windows. Disambiguated with an
explicit uint64(...) cast, matching the pattern already used at
AchievementMgr.cpp/CalendarHandler.cpp/PlayerTalent.cpp. This is the
Docker/GCC build's first real --clean-first pass since these test files
were added, per the plan's own prediction that removing libace-dev from
the container would surface latent gaps immediately.

Zero-ACE grep gate (`grep -rlE '\bACE_[A-Z]|#include <ace/' src/`):
8 files, all comment-only historical citations, zero live code --
mop_auth_test.cpp, mangosd.cpp, Service.h, Bots/CMakeLists.txt,
proto/ClientConnection.cpp, proto/CMakeLists.txt, Common.h, Util.cpp.

Verified: MSVC 2022 Release ALL_BUILD is green (zero errors, only
pre-existing/unrelated warnings in WheatyExceptionReport.cpp and third-
party dep/ code). The m4build GCC 13/Ubuntu 24.04 container had
libace-dev/libace-7.1.2t64 purged before this build -- configure and
`make -j16` are green from a bare CMakeCache. ctest on both toolchains:
mop_auth/mop_framing pass, the same 20 pre-existing mop_*_source
failures recur unchanged, no new failures.

Boot smoke against the existing server_install_four install: realmd
listens on 3724 (16 IOCP workers); mangosd reaches "World Initialization
Complete", listens on 8085 (16 IOCP workers), and a SIGTERM runs the new
Master/ShutdownWorld teardown cleanly through KickAll, final
UpdateSessions, StopNetwork, MapUpdater deactivation, and DB delay
thread halt, exiting "Bye!". No DB version gap this time (CP3 hit
World db v21 vs required v23 on this same install; it has since been
migrated externally, unrelated to this commit).

+++ src/game/Server/tests/mop_movement_packets_test.cpp
@@ -649,7 +649,7 @@ static void test_complementary_and_empty_state()
 static void test_server_built_embedded_guid()
 {
     MovementInfo info;
-    info.SetMoverGuid(ObjectGuid(0x8070605040302010ull));
+    info.SetMoverGuid(ObjectGuid(uint64(0x8070605040302010ull)));
     RefState state;
     state.guid = 0x8070605040302010ull;
     state.flags = state.flags2 = state.timestamp = 0;
@@ -779,7 +779,7 @@ static void test_transport_stop_monster_move_fixture()
 {
     WorldPacket packet(SMSG_MONSTER_MOVE, 64);
     Movement::PacketBuilder::WriteStopMovement(G3D::Vector3(7.0f, 8.0f, 9.0f), 0xAABBCCDDu,
-        packet, ObjectGuid(0x8877665544332211ull), ObjectGuid(0xA8A7A6A5A4A3A2A1ull), int8(3));
+        packet, ObjectGuid(uint64(0x8877665544332211ull)), ObjectGuid(uint64(0xA8A7A6A5A4A3A2A1ull)), int8(3));
 
     std::vector<uint8> const expected {
         0x00, 0x00, 0x10, 0x41, 0x00, 0x00, 0xE0, 0x40, 0xDD, 0xCC, 0xBB, 0xAA,
@@ -796,7 +796,7 @@ static void test_linear_monster_move_fixture()
     Movement::MonsterMoveData move;
     move.position = G3D::Vector3(1.0f, 2.0f, 3.0f);
     move.splineId = 0x11223344u;
-    move.moverGuid = ObjectGuid(0x0807060504030201ull);
+    move.moverGuid = ObjectGuid(uint64(0x0807060504030201ull));
     move.duration = 1000;
     move.compressedPath.push_back(G3D::Vector3(1.0f, 1.0f, 1.0f));
     move.uncompressedPath.push_back(G3D::Vector3(9.0f, 10.0f, 11.0f));
